docker_dowlog/requirements.txt
advanced_alchemy==0.26.2
aiofiles==24.1.0
aiogram==3.15.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.2
alembic==1.14.0
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
async-timeout==5.0.1
asyncpg==0.30.0
attrs==24.3.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
docker==7.1.0
EditorConfig==0.17.0
exceptiongroup==1.2.2
Faker==33.1.0
fast-query-parsers==1.0.3
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.4
jsbeautifier==1.15.1
litestar==2.13.0
litestar-htmx==0.4.1
loguru==0.7.3
magic-filter==1.0.12
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msgspec==0.18.6
multidict==6.1.0
polyfactory==2.18.1
propcache==0.2.1
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
rich-click==1.8.5
schedule==1.2.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.3
websockets==14.1
yarl==1.18.3
